+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/base/callbacks/OnMediaGroup.kt
@@ -3,38 +3,32 @@ package com.github.insanusmokrassar.AutoPostTelegramBot.plugins.base.callbacks
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.database.PostTransaction
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.models.PostMessage
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.Plugin
+import com.github.insanusmokrassar.AutoPostTelegramBot.checkedMediaGroupsFlow
 import com.github.insanusmokrassar.AutoPostTelegramBot.mediaGroupsListener
 import com.github.insanusmokrassar.AutoPostTelegramBot.plugins.base.commands.usersTransactions
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.subscribe
+import com.github.insanusmokrassar.AutoPostTelegramBot.utils.flow.collectWithErrors
 import com.github.insanusmokrassar.TelegramBotAPI.types.ChatIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.FromUserMessage
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.MediaGroupMessage
+import kotlinx.coroutines.*
 import java.util.logging.Logger
 
 private val logger = Logger.getLogger(Plugin::class.java.simpleName)
 
-class OnMediaGroup(
+internal fun CoroutineScope.enableOnMediaGroupsCallback(
     sourceChatId: ChatIdentifier
-) {
-    init {
-        mediaGroupsListener.subscribe(
-            {
-                logger.throwing(
-                    OnMediaGroup::class.java.canonicalName,
-                    "Perform message",
-                    it
-                )
-                true
-            }
-        ) {
-            invoke(it.data, sourceChatId)
+): Job = launch {
+    checkedMediaGroupsFlow.collectWithErrors(
+        { update, e ->
+            logger.throwing(
+                "Media groups AutoPost callback",
+                "Perform update: $update",
+                e
+            )
         }
-    }
-
-    private fun invoke(
-        messages: List<MediaGroupMessage>,
-        sourceChatId: ChatIdentifier
     ) {
+        val messages = it.data
         val first = messages.first()
         if (first.chat.id == sourceChatId) {
             val id = when(first) {